parallelize translation download requests
Downloading translations from Transifex is slow. For our project with the most translations to pull, it was taking about 20 minutes. By parallelizing the download requests, that number goes down to about 25 seconds.

If we wanted to get that even faster in the future, we could parallelize the work on the Transifex resources as well, rather than just the translations within a resource. (Because right now the baseline time is still multiplied by the number of resources a project references; for example, our Android client uses at least 4.)

Because this change increases our rate of requests, it risks us hitting Transifex rate limits. We'll address that if/when the time comes. (See the comments in the code for possible approaches.)